requirements.txt
# Python 3.9+.
streamlit
pandas
google-genai
python-dotenv